Make global loggers safe for concurrent use
I'd originally intended zap's global loggers to be a consenting-adults sort of
API in which easily-avoided data races are okay. After just a few migrations,
it's become clear just how naive that plan was.

This PR adds tests to verify that the global loggers are safe to use and
reconfigure concurrently, and it converts the global variables to (short)
functions.